Update folder structure and ReadMe
These changes are to reflect the structure of the Dojo. We will not
follow one book specifically, we will initially start with Elixir In
Action and potentially move onto others in the future.
Katas folder has been added so when we haven't got many member attending
we can do a Kata instead of going over the book.